Okii so...
I've managed to move the "about us" link to the bottom of the navbar using vh instead, and it remains responsive.

I fixed the navbar icons growing bigger and smaller suddenly, whenever the navbar is expanded. I just removed the "min-width" elsewhere in the CSS.

I wrapped the navbar buttons in <a> tags within the <li> tags. Turns out, it was unnecessary. But it's done now, so...

The images on the cards still sometime overlap the navbar when it expands. The cursor has to pass over the card image to get to the navbar. I should probably do something about that.